Move all 'tasks' tests to test_tasks module
